+++ app/Http/Controllers/API/Admin/Permissions/PermissionsController.php
@@ -19,29 +19,32 @@ class PermissionsController extends Controller
 
     public function index(Request $request)
     {
-
+        $this->authorize('permission_access');
         return $this->permissionsService->getAllPermissions($request);
     }
 
     public function store(StorePermissionRequest $request)
     {
 
+        $this->authorize('permission_create');
         return $this->permissionsService->store($request);
     }
 
     public function edit($id)
     {
-
+        $this->authorize('permission_edit');
         return $this->permissionsService->edit($id);
     }
 
     public function update(UpdatePermissionRequest $request, int $id)
     {
+        $this->authorize('permission_edit');
         return $this->permissionsService->update($request, $id);
     }
 
     public function destroy($id)
     {
+        $this->authorize('permission_delete');
         return $this->permissionsService->destroy($id);
     }
 }
